### 2021-03-15 #50 @rsankar9
[Re] A Reservoir Computing Model of Reward-Modulated Motor Learning and Automaticity
**Original article:** A Reservoir Computing Model of Reward-Modulated Motor Learning and Automaticity

**Scientific domain:** Computational Neuroscience
**Programming language:** Python, MATLAB
