### 2021-03-22 #50 @schmidDan
@benoit-girard if you don't mind, I'd be happy to jump in reviewing this submission. The interesction of recurrent networks and reward-modulated/reinforcement learning is matching my interests and fluidity with Python (/Matlab) is given as well.
