### 2021-03-25 #50 @benoit-girard
@ChristophMetzner Thanks for answering!
@stephanmg We have no general policy (so far) at ReScience C. Do you think you could do it within 3 weeks? i.e. aiming at the 15th of April.
