### 2021-04-06 #50 @rgutzen
@benoit-girard @rsankar9 Here is my review. I hope you find it helpful, and I'm happy to discuss if any points are unclear.

Overall, this is a very clearly structured paper and replication. Especially the investigation of the model limitations is very valuable. It's easy to read and to understand. The reimplementation is understandable and runs out of the box. I have some comments, mainly concerning how the agreement of the reproduction/replication is evaluated and presented.

## Replication/Reproduction
* It is great that you perform both a reproduction with the original code and a replication with your own implementation and separate clearly between them. Just in the discussion, their evaluation could be described more precisely.
* The main issue I see in the submission, is the lack of explicit criteria used to judge how successful the reproduction and replication are. You use vague qualitative assessments such as "successfully closely", "function as presented in the paper", and "model does reproduce the results", however it is unclear how you arrive at this evaluation and what this entails in detail. Shouldn't the traces in the paper be reproduced exactly by using the same Matlab code and seed? What features of the trace deviation are relevant or just the average deviation from the target? Are some deviations from the target trace more or less worse than others? You should preferably aim for a quantitative evaluation of the similarity of traces that incorporates the relevant performance measures of the model, or if you provide a qualitative evaluation 'by eye', it is helpful to explain which features of the model behavior you focus on. In the figures, also a 'distance from target' measure as in the original paper is not shown, which would be already very helpful in that regard.
* Using such a more quantitative evaluation the comparison of paper, Matlab, and Python version (+ modified) could be more differentiated. For example, from the figures alone the Python version seems to have more variation (be less stable/ more chaotic) than the Matlab version, nearly throughout the tasks (especially 1&2). This is however not addressed or explained.
* Another point, that is mentioned but apparently not really factored into the evaluation of the model is its dependency on the initial seed. A model that deviates from its intended behavior in 50% of cases just because of the choice of seeds seems jarring, especially as it is not reported in the original paper. It lets the reader wonder to what degree the original results were accidental and cherry-picked. To better grasp the influence of the seed and robustness of the model I'm missing the figures of the runs with different seeds in the paper (e.g. as supplements) and in the best case a quantification of the variation due to the chosen seed.

## Text
* It would be helpful for the reader to state the level of reproduction and replication already early in the intro. Something in the realm of 'successful with minor/moderate limitations and modifications'
* missing an explanation of 'e' in equation 2
* "verify performance of the model". This formulation is somewhat confusing to me. Evaluating a model's output against another would rather be a validation instead of a verification, and performance may also refer to the usage of computational resources instead of accuracy. I would suggest a reformulation explaining more precisely what you are doing e.g. evaluate the accuracy/robustness/behavior of the model / validate the model against its Python re-implementation / verify whether the original script produces the published results.
* 2.1. l.7: "Post the training phase" When is training stopped? Is the training time fixed? What are the criteria?
* Figure 4. is not referenced in the text
* Where the agreement of the model outputs is described as "successfully closely" or similar, a more precise statement would be suited based on prior defined criteria (as described above).
* Similarly, the discussion needs to be more differentiated and transparent in how you come to your conclusions e.g. "For Task 1 and 2, we can now confirm that the three algorithms function as presented in the paper. For Task 3, the SUPERTREX model’s behaviour is also reproducible, ..."

### Typos/Grammar/Suggestions
* l.5 fail -> fails
* l.6 utilise -> use
* In the first sentence of the task descriptions, e.g. "Here, we compare the simulations of the author scripts and our re‐implementation for Task 1 using FORCE, RMHL and SUPERTREX, with ..." last comma is not needed.
* to initialise -> initializing in "To do so, we run the simulations with the default seed and
repeat it ten times with different seeds to initialise the random number generator." otherwise it may sound like you run the simulations to initialize the RNG.
* demarcates -> marks the separation of
* in 3. Modification: use numbered bullets instead of "* One, ..."
* in 4. Discussion, l.3: out -> our

## Figures
* This might not be technically possible, but it would help with the visual inspection of the butterfly traces if all three instances (Matlab, original, Python) would have more similar linewidths and colors (and same butterfly height/width ratio).
* As already mentioned above, the figures would benefit from also including a distance-from-target trace, similar to the MSE in figure 5.
* Why don't you also show the time series from the paper for comparison?
* Fig. 2 caption: "shows the actual output". Aren't both rows the actual output just in different representations? Maybe reformulate as 'time series.
* Fig. 3, a) & c): missing time series. Is there a particular reason for that?
* Fig. 5: Scale of MSE is very hard to read.
* And again as a reader I would be very interested to also see the runs with other seeds or some aggregate evaluation of them.

## Code
* Regarding the Code and the repository, there is very little to object from my side. It is well structured, documented, easy to reuse, and reproduces all figures.
* I wouldn't mind having a requirements.txt or environment.yaml file instead of the specification in the Readme.
* The ten arbitrary seeds for the RNG are chosen separately for the different models, so the comparisons are both between models and seeds when they are not fixed (which can have a major influence, as you showed). This, however, doesn't become clear from the description in the paper.
* Just for curiosity, why did you choose a custom Python implementation of the model, what kept you from using a simulator engine like Brian, Nest, or even PyTorch?
