### 2021-04-15 #50 @schmidDan
@rsankar9 I enjoyed reading and reviewing your submission. Manuscript and code are well written! I had some things to point out in both, code and manuscript. I'd consider them as minor improvements to the quality of the contribution, as the validity of your main claims are assured nevertheless. Please find my review below @rsankar9 @benoit-girard .

## Review Summary

* Full replication | Partial replication | Failed replication | Reproduction: I believe it's a "Partial replication", but I have two things to note here @benoit-girard:
  * The authors mention in their manuscript's section "2 Comparison with Python Re-implementation" that they had access to the Matlab code of the original publication's authors. This essentially introduces a bias to the replication, which became evident by that the authors made modifications to their Python code based on how the Matlab code differs from methods described in the original publication. So I believe without the Matlab code at hand it would've been a "Failed replication" due to the seemingly missing information in the original publication. The authors consider their partial replication being a "Re-implementation", which sort of reflects this circumstance and very well point out the differences they discovered between the original paper and the Matlab code. I would consider it therefore a successful partial replication, but wanted to point out this circumstance nevertheless.
  * Is there a way to denote the difference in replications between full and partial in the article's tag (i.e. [Re] vs. [Re\\] or something)? As so far the manucsript is tagged as a "Replication" besides replicating not all experiments (i.e. "Partial replication").
* Licensing: BSD-2-clause license for both Matlab code (by Pyle and Roberts) and Python code (by the contributors of this submission)
* Reproducibility of the replication: I was able to run the Python code without any problems. The Matlab code did not work under GNU Octave 6.2.0. As Matlab is closed-source I'm not sure what the directive would be here either way @benoit-girard?
* Clarity of the code: Overall well structured and documented code.
* Clarity and completeness of the accompanying article: Concise and easy to read, captures the main claims and contributions of the original publication and points out robustness issues encountered in certain experiments as well as a possible solution. Some points could be alittle bit clearer (see my comments below).

## Remarks w.r.t. the manuscript

* Layout: reading through the references is a little bit complicated, since LaTeX obviously cluttered some graphics inbetween the list of references (Figs. 5, 6). The figures should be placed before the "References" section starts.
* References: Reference 1 is missing the journal name and is formatted differently than refs 2-6 (title in bold instead of journal, no quotation marks around title etc.). Some journal names are in title case others in sentence case.
* Framework:
  * "This allows for partially unsupervised learning" (p.2) - what is  meant by "partially unspervised learning"?
  * The article should make for a self-contained read without reiterating every detail from the original publication. For my taste, this would include stating the complete model equations, i.e. reservior equations, computation of `z_1`, `z_2` from `r`, and mentioning the numerical solver used to simulate the system of differential equations, so that the reader knows about the meaning of the weights updated by eq.1-3.
  * While FORCE and RMHL acronyms are epxplained, this was omitted for SUPERTREX (p.2).
* Task:
  * "each tasks uses [...] multi-segmented arm" (p.2) - some tasks are based solely on the pen's position. Admittedly, one could view that as a 0-segment arm condition, though.
  * Concise language: "A non-linear inverse transformation would be required" (p.2) - transformation of what? having read the original work this is clear to me, but might not be clear for someone who doesn't know/forgot about the details in the original work.
  * Reading this section as well as "Task 1" and "Task 2" it might seem to the reader as if RMHL and SUPERTREX are learned with non-scalar error values for Task 1, as for Task 1 it is solely noted that "known target" output is used for training while for Task 2 it is explicitely stated that the error is a scalar. Indeed, for Task 1 the error (to the reward-modulated learner) is as well of scalar type. Critically, though the distinction comes from how the scalar is computed: output and target domain match in Task 1 - distance of the pen's position is indeed the target value of the output - the error couples linearly to the correct solution's output values -, while for Task 2 and 3 the output domain is angular values, but the target domain is the end effector's position - the error is non-linearly coupled to the correct solution's output values.
* Task 1,2,3:
  * The explanation should note, that during testing the feedback is provided via teacher forcing (cf. the original publication), since it was shown to be of quite important impact for the model performance.
* Task 3:
  * "We compare the performance of the MATLAB scripts and our Python adaptation for the three algorithms on Task 2, with the results presented in the article." - This sentence isn't clear to me. I believe it should say "Task 3" instead of "Task 2" here? Also describing your Python (re-)implementation as an "adaptation" might be a little bit confusing in the light that you later on indeed modify (i.e. adapt) your Python implementation to increase robustness.
  * I don't believe the text accurately reflects the degree to which the results are in line with the original publication here. While the later sentences acknowledge a deviation from the target contour, the summarizing introduction ("The MATLAB scripts and the Python re‐implementation are able to successfully reproduce the results presented in the paper with the default seed as well as with the 10 arbitrary seeds for the RMHL algorithm.") raises the impression, that the results in the orginial publication look about the same. Indeed, they don't, which is also why you changed the seeds (which nevertheless do not make up for the complete discrepancy w.r.t. the original publication). See also my comments w.r.t. "Results" in "Remarks w.r.t. the implementation".
* Modification:
  * Figure 4 is never referenced, but I feel like it should be noted somewhere within this section?
  * "We observe that RMHL performance is comparable to the original Task 2, [...]" - this is nowhere shown, right? So maybe add a "(not shown)".
  * The "two minor alterations" you stated refer to Tasks 1-3 throughout the section. To me it is not entirely clear whether the results of the privous sections have been generated with or without these alterations. From the structure of the manuscript and the labelling of the figures I would expect that the alterations mentioned are only applied for experiments conducted within "3 Modification". But, on the other hand, why would one state then how these alterations relate to Tasks 1-3?
  * It is great to have the two alterations stated as a bullet point list. What might be a little confusing is the reiteration of the two points (with an added interpretation of them) in the subsequent paragraph as by using "also" in "We also increase the error threshold [...]" it seems, on a first look, like here comes an additional alteration on top of the ones from the bullet point list, while it is indeed the second item of the list.
  * "[...] the model is able to perform well with up to 50 arm segments (Figure 5)." - Again, I believe "well" is a little bit optimistic for some of the results. For e.g. 10 arm segments (Fig. 5) I wouldn't use the word "well", while for e.g. 6 arm segments this seems to be justified.
* Figures:
  * No indication for the temporal extent of the plots is given (x-axis) (cf. original publication).
  * I believe the "Original" butterfly plots are taken from the original publication. If so, this needs to be stated (and potential permission issues would have to be clarified with the holder of their copyrights).
  * The "Original" butterfly plot looks squeezed w.r.t. the horizontal extent. Is this just a visual feat, or as well corresponding to differencees in the underlying implementation?
  * For viewing the document digitally, it might be beneficial to use vector graphic plots instead of pixelated information. Note the difference when zooming in into the figures. Just in case this is easily realizable and doesn't require too much effort.
  * The different plots within a figure aren't drawn to the same scale (not the difference in "MATLAB" and "Python" red butterfly outlines, or the vertical scaling of e.g. x and y axes), which makes it harder to visually judge relative perofrmance. Same goes for the horizontal scaling of temporal axes where weight changes are depcited (i.e. the gray vertical bars don't align with them of the angular evolutions).
  * No plots for the evolution of the performance metric (distance from target) and weights (norm of the weight matrix) are shown. This would aid the comparison with the original publication and, I believe, would make the case for a successful replication even stronger.
  * Some plots (in Figs. 4,5) have scales for their vertical axis plotted, which is great. Unfortunately, the ticks of these scales are too small and squeezed to be interpretable.
  * For some of the temporal plots a slighlty increased linewidth would helpful.
  * Up to taste: Using the same color tone (same blue) for "MATLAB" and "Python" temporal evolution plots would be pleasant.
  * Figure 1: It seems like "MATLAB" and "Original" butterfly plots are consisiting of one period of drawing the shape (blue line), while "Python" seems to have multiple periods visualized. Is this the case, or are the different periods just aligning perfectly?
  * Figure 2:
    * For the plots from the original article there seems to be some vertical gray artifact line to the left of each plot (it vanishes and appears w.r.t. my reader's zoom factor).
    * The caption seems to have a copy-paste-error: ""The second row shows [...] (x and y coordinates, in this case) [...]" - actually, the angles are depicted (according to the vertical axes' tag).
  * Figure 3:
    * For the plots from the original article there seems to be some vertical gray artifact line to the left of each plot (it vanishes and appears w.r.t. my reader's zoom factor).
    * Suplot "(a)"/"(c)": "The target time-series is imitated well by the model during the training phase [...]" - this information is not shown and should be denoted as such (i.e. "(not shown)").
    * Subplot "(b)": The y-axes of the plots of evolutions of the angles are somehow having an additional, squashed tag for each angle on to their left. These are probably due to how the plots have been generated.
    * Subplot "(c)": I'm not sure whether "[...] with slight divergences [...]" does the subplots do justice. Qualitatively, the results are still of butterfly shape, but quantitatively, I believe, the performance metric will be quite impacted during some phases of each period (not shown).
  * Figure 4:
    * The differently colored lines of the evolution of the matrices' norm isn't described (i.e. green = exploration, purple = mastery pathways / `W_1` and `W_2`).
* Discussion:
  * "Only two necessary details were missing, [...]" - if they have been provided by the original authors upon request, it would be worth to mention it. Or have they been found by trial-and-error?
  * "[...] the inclusion of a crucial learning rate of 0.0005 [...]" - can you explain which learning rate this refers to? If I read the original publication's method section correctly, then they state a learning rate of k=0.5 for Tasks 1 and 2 and k=0.9 for Task 3 (p. 1457).
  * "[...] and a compensatory factor of 0.5 within the update of the readout weights of the exploratory pathway [...]" - I'm not sure what this exactly means wihtout an equation. I believe it's just another mulitplicative factor to the update. If so, why wouldn't one state it as part of the learning rate then? I.e. "[...] the inclusion of  a crucial learning rate of 0.0005 for Task 1 and 0.00025 for Tasks 2 and 3". [edit: ...reading your code as well I understood that you treat it as a separate factor, since it is encapsulated in a "compensation" method.]
  * "For Task 3, the SUPERTREX model's behaviour is also reproducible, [...]" - question to clarify: reproducible or replicable?
* Smaller comments:
  * fully supervised vs. fully-supervised: inconsistent usage of a hyphen throughout the text.
  * Inconsistend usage of language: sometimes you're referring to the Python code as "re-implementation" sometimes as "adaptation". Or are these two referring to different parts of your code? If yes, which ones?
  * Discussion: "provided by the author" should probably be plural ("authors") or "by the corresponding author"?
  * Task 1: "The article claims that :" - a whitespace between "that" and ":".
  * Task 3: "The article claim that" - should be "claims", ":" missing at the end.
  * Discussion: "[...] with out modular and user-friendly Python replication." - should be "our".
  * Discussion: "[...] for task 3" - throughout the text "Task ..." was used as a proper noun, so it should be "Task 3" here as well.

## Remarks w.r.t. the implementation

* I understood that the Matlab implementation was provided by the authors of the original publication. But did they as well agree to have it hosted on Github as well as being considered part of your submission to ReScience C?
* `README.md`:
  * The file mentions "The Python re-implementation has been submitted to ReScience C", but if I understand correctly I need to review the Matlab part as well to check for whether the results in your manuscript hold. So, if the Matlab part has to be part of my review, isn't it then as well part of the ReScience C submission? And if so, as it is provided by the original authors, how does this relate to tagging this submission as Replication vs Reproduction (@benoit-girard)?
* Requirements are stated in e.g. `Code/Python implementation/Reimplementation/README.md`. I would welcome it having them as a proper `requirements.txt` (with versions enforced by "==") in the respective folders, as it would be the common way to tell a Python user which packages to install.
* Comparison of code with original publication:
  * Butterfly generator functions: Comparing with the original publication (Sec. 4.5 Tasks) it seems like there is a discrepancy in what was described and what is implemented in Matlab (respectively re-implemented in Python). They say `x(t) = r(t)*cos(t)` with `t` in 0 to 10^4 ms, while every generator method I checked in the implementation does `x(theta) = r(theta)*cos(theta)` with `theta` in 0 to 2pi. Maybe that would be worth a comment in the code or similar.
  * For the psi function (Sec. 4.5 Tasks) it is written that `Psi(x) = 0.025 * (10.0 * x)^0.25` is used for Task 3, while in the code (Matlab and Python) it is set to `Psi(x) = 0.005 * (10.0 * x)^0.25`. Maybe it would be good to point that out.
  * The distance from target was said to be computed from "sqrt(\bar{MSE})" (p. 1455 of the original publication). Comparing with the Python code (e.g. ModelFORCE.py:272,379) the "sqrt" operation seems to be commented out(?).
  * Smoothing time constant for `z` (\bar{z}) seems to be 10 according to the paper (p. 1455), while in the Python parameter files (e.g. "simulation_parameter_file_Task1_FORCE.json") `tau_z: 1`.
  * Comments: At some part in the code there are comments like "can be removed" (e.g. ModelRMHL.py:165,174) or "Possibly wrong to use for" (ModelSUPERTREX.py:230) or unused variables like "unnec" in e.g. ModelSUPERTREX.py:117, or "smooth noise?" in Task2_ST_Seg2.py:134, which overwrites the result from Task2_ST_Seg2.py:133. Or "weird stuff in code" in Task.py:132. Are all/some of them are meant to make it into the published version of the code, or should they be removed beforhand?
* Results: While the results are as described in the accompanying article, taking a look at all the results in the different "Results" folders, it seems like the Matlab implementation yields higher performance (lower MSE) comared to the Python implementation (e.g. comparing the "MSE.png" files for "FORCE_Task1" it seems like Matlab reaches as low as 10^-6, while python reaches just 10^-4). Do you have any idea why?
* The way of initialization "J" (e.g. ModelFORCE.py:106-113) seems a little bit complicated. I'm just wondering whether [`scipy.sparse.random`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.random.html#scipy.sparse.random) wouldn't be an alternative (with using the respective `data_rvs` argument)?
