### 2021-06-17 #50 @benoit-girard
@schmidDan sorry for being so late to answer your questions:

- I don't think that we have a title marker chosen for partial replications.
- concerning the code, if I understand well, the code that is really submitted here as a replication is the python one. If so, it is the only one expected to work and to be reviewed.
