### 2021-06-27 #50 @rsankar9
Thank you for the very thorough and constructive reviews. I do apologise for my delayed response. I have made several changes to the manuscript, based on your reviews.

- The plots have been made more clear.
- Two metrics have been added to quantify the divergence. This has been both tabulated and indicated in the text and plots.
- A terminology subsection has been added to clarify the terms used for each implementation.

More specific questions have been addressed inline for each review.
The code and manuscript has been updated in the repository.

@benoit-girard : I do have the permission of the authors to publish their MATLAB code here. However, I do not have the permission to use the images from their paper (the trajeectory traces). I shall contact them regarding this and keep you posted.
