### 2021-06-27 #50 @rsankar9
@rgutzen : Thank you for your review! Please find below my responses to your comments (inline).

----

> @benoit-girard @rsankar9 Here is my review. I hope you find it helpful, and I'm happy to discuss if any points are unclear.
>
> Overall, this is a very clearly structured paper and replication. Especially the investigation of the model limitations is very valuable. It's easy to read and to understand. The reimplementation is understandable and runs out of the box. I have some comments, mainly concerning how the agreement of the reproduction/replication is evaluated and presented.
>
> ## Replication/Reproduction
>
>     * It is great that you perform both a reproduction with the original code and a replication with your own implementation and separate clearly between them. Just in the discussion, their evaluation could be described more precisely.
>
>     * The main issue I see in the submission, is the lack of explicit criteria used to judge how successful the reproduction and replication are. You use vague qualitative assessments such as "successfully closely", "function as presented in the paper", and "model does reproduce the results", however it is unclear how you arrive at this evaluation and what this entails in detail. What features of the trace deviation are relevant or just the average deviation from the target? Are some deviations from the target trace more or less worse than others? You should preferably aim for a quantitative evaluation of the similarity of traces that incorporates the relevant performance measures of the model, or if you provide a qualitative evaluation 'by eye', it is helpful to explain which features of the model behavior you focus on. In the figures, also a 'distance from target' measure as in the original paper is not shown, which would be already very helpful in that regard.

* 	This is a very valid point and I understand the concern regarding the evaluation criteria. The replication was considered successful by visually inspecting the plot of the MSE of the trajectory, i.e. the deviation of the simulated trajectory from the target trajectory. To address this in a more quantitative manner, I have included two metrics in the manuscript now.
	*  One, the mean deviation. This measures the average deviation ("Distance from target") over the test phase, for each simulation, and then, presents a summary statistic over 11 simulations with different seeds for the random number generator.
	*  Two, I set a threshold on this metric to categorise the model performance as satisfactory or unsatisfactory. I present the proportion of simulations, of each task, for which the performance was satisfactory. Moreover, I mention the proportion of simulations for which the model fails to proceed to completion (when weights increase exponentially).
*  Hopefully, these two criteria should improve the quantitative assessment of the performance of the different implementations.
*  I've supplemented the vauge terms "succesful closely" with these two metrics, and substituted them with satisfactory/unsatisfactory wherever applicable.
*  I've included the MSE plots for the simulations, along with the mean deviation over the test phase. This should address if some anomalies in the test phase are better or worse than others.
*
>	Shouldn't the traces in the paper be reproduced exactly by using the same Matlab code and seed?

* 	This is a valid point which we had also initially expected. However, this is not the case due to the differences between the random number generators in MATLAB and Python. While we were able to replicate the exact same behaviour for most MATLAB functions, we were unable to do so for one particular function, ``sprandn'' (also mentioned in the paper). This results in the initialisation of the J matrix being different and hence, in the differences in the final trajectories. To address this query better, I've included the results of one simulation, where I've plugged the J matrix produced in MATLAB into the Python adaptation, with the same random seed. You will see that in doing so, we obtain the exact same results.

>     * Using such a more quantitative evaluation the comparison of paper, Matlab, and Python version (+ modified) could be more differentiated. For example, from the figures alone the Python version seems to have more variation (be less stable/ more chaotic) than the Matlab version, nearly throughout the tasks (especially 1&2). This is however not addressed or explained.

* This discrepancy was mainly due to an error on my part. Task 1 and Task 2-3 use teacher forcing in the test phase in two different ways. I had implemented them using the same method for Task 1-3. This led to the more chaotic results in Python. I've rectified that now, and verified that the adaptation is accurate.

>     * Another point, that is mentioned but apparently not really factored into the evaluation of the model is its dependency on the initial seed. A model that deviates from its intended behavior in 50% of cases just because of the choice of seeds seems jarring, especially as it is not reported in the original paper. It lets the reader wonder to what degree the original results were accidental and cherry-picked. To better grasp the influence of the seed and robustness of the model I'm missing the figures of the runs with different seeds in the paper (e.g. as supplements) and in the best case a quantification of the variation due to the chosen seed.

-  We do take into account the dependency of the initial seed. Even then, it is difficult to judge if the results were cherry-picked, as we are not aware of the initial seeds used by the authors. We mention that we assume that the default seed would have been used by the authors (due to lack of any mention to the contrary in their code), and make further inferences on this basis. To quantify this here, I've used the summary statistic of the mean deviation over 11 seeds, and the proportion of seeds with satisfactory performance. This should hopefully shed light on the extent of the dependance on the seed.

> ## Text
>
>     * It would be helpful for the reader to state the level of reproduction and replication already early in the intro. Something in the realm of 'successful with minor/moderate limitations and modifications'
>
>     * missing an explanation of 'e' in equation 2
>
>     * "verify performance of the model". This formulation is somewhat confusing to me. Evaluating a model's output against another would rather be a validation instead of a verification, and performance may also refer to the usage of computational resources instead of accuracy. I would suggest a reformulation explaining more precisely what you are doing e.g. evaluate the accuracy/robustness/behavior of the model / validate the model against its Python re-implementation / verify whether the original script produces the published results.
>
>     * 2.1. l.7: "Post the training phase" When is training stopped? Is the training time fixed? What are the criteria?
>
>     * Figure 4. is not referenced in the text
>
>     * Where the agreement of the model outputs is described as "successfully closely" or similar, a more precise statement would be suited based on prior defined criteria (as described above).
>
>     * Similarly, the discussion needs to be more differentiated and transparent in how you come to your conclusions e.g. "For Task 1 and 2, we can now confirm that the three algorithms function as presented in the paper. For Task 3, the SUPERTREX model’s behaviour is also reproducible, ..."

* I've included the state of the level or replication in the intro. "We were able to successfully reproduce the model performance in Python, for two tasks out of the three presented in the original article. For the third task, we were able to do so with limited robustness. We address this by introducing some modifications, and discuss how their inclusion vastly improves the robustness as well as scalability of the model."
* Improved the explanation for equation variables.
* I've replaced the usage of "verify" wherever not applicable. For e.g. "To validate the re-implementation". Regarding the aim of the manuscript, it attempts to both verify the results presented in the original article, as well as validate the Python re-implementation. It, therefore, continues to use the word "verify" where applicable. This distinction should be more clear in the manuscript now.
* I've included more information about this in Section 1.2. I've used the same protocol presented in the original article. Yes, the training time is fixed.
* I've included the reference to Fig 4.
* Using the two metrics, a more precise statement has been included wherever applicable.
* Our inference should now be more transparent with the help of the metrics for deviation and proportion of satisfactory runs, for accuracy and robustness, respectively. To be more concise, I've edited the conclusion to "We conclude that the results presented in the paper are reproducible using the original scripts provided by the authors, and replicable in Python with comparable performance."

>
> ### Typos/Grammar/Suggestions
>
>     * l.5 fail -> fails
>
>     * l.6 utilise -> use
>
>     * In the first sentence of the task descriptions, e.g. "Here, we compare the simulations of the author scripts and our re‐implementation for Task 1 using FORCE, RMHL and SUPERTREX, with ..." last comma is not needed.
>
>     * to initialise -> initializing in "To do so, we run the simulations with the default seed and
>       repeat it ten times with different seeds to initialise the random number generator." otherwise it may sound like you run the simulations to initialize the RNG.
>
>     * demarcates -> marks the separation of
>
>     * in 3. Modification: use numbered bullets instead of "* One, ..."
>
>     * in 4. Discussion, l.3: out -> our

* I believe it should be fail.
* Edited.
* Edited to: "Here, we compare the simulations of the original scripts and our adaptation for Task 1, using FORCE, RMHL and SUPERTREX, with the results presented in the article. "
* Edited.
* Edited.
* Edited.
* Edited.

> ## Figures
>
>     * This might not be technically possible, but it would help with the visual inspection of the butterfly traces if all three instances (Matlab, original, Python) would have more similar linewidths and colors (and same butterfly height/width ratio).

* Fixed.

>
>     * As already mentioned above, the figures would benefit from also including a distance-from-target trace, similar to the MSE in figure 5.

* Fixed.

>
>     * Why don't you also show the time series from the paper for comparison?

* If I include them, the plots would be considerably scaled down, and this would impact readability. And since the MATLAB results should be the same, in all cases except Task 3, it doesn't add any extra information. In Task 3, since we were unable to reproduce their results, there seems to be no reason to include it for comparison. Also, I wish to limit the no. of figures we re-use from the original article.

>     * Fig. 2 caption: "shows the actual output". Aren't both rows the actual output just in different representations? Maybe reformulate as 'time series.

* Edited across the manuscript.

>     * Fig. 3, a) & c): missing time series. Is there a particular reason for that?

* Yes. I chose not to add these plots, at the cost of the clarity of Fig. 3b, as I believe in these 2 cases, it does not provide any additional information. It can be infered clearly from the previous 2 figures and the figures in the original paper.

>     * Fig. 5: Scale of MSE is very hard to read.

* Fixed.

>     * And again as a reader I would be very interested to also see the runs with other seeds or some aggregate evaluation of them.

* I've included an aggregate evaluation using the proportion over 11 simulations. I've also included the aggregate deviation metric in a table. The simulations with the other seeds are available in the repository in the 'Results' folder for each variant.

>
> ## Code
>
>     * Regarding the Code and the repository, there is very little to object from my side. It is well structured, documented, easy to reuse, and reproduces all figures.
>
>     * I wouldn't mind having a requirements.txt or environment.yaml file instead of the specification in the Readme.

Included a requirements file.

>     * The ten arbitrary seeds for the RNG are chosen separately for the different models, so the comparisons are both between models and seeds when they are not fixed (which can have a major influence, as you showed). This, however, doesn't become clear from the description in the paper.

* Included a clarifaction: "Except for the default seed, the ten arbitrary seeds are different for the Python and MATLAB simulations, and for each algorithm-task combination. "

>     * Just for curiosity, why did you choose a custom Python implementation of the model, what kept you from using a simulator engine like Brian, Nest, or even PyTorch?

* Brain and Nest are simulators used generally for spiking neural networks. This is a rate-coded model, hence, it would be unnecessary. Also, the scientific stack in Python, including numpy, scipy, etc, is fully sufficient for a simple one or two layer reservoir computing framework. This removes the need for a more complex tool.
