### 2021-06-27 #50 @rsankar9
@schmidDan: Thank you for your review! Please find below my responses to your comments (inline).

----

> @rsankar9 I enjoyed reading and reviewing your submission. Manuscript and code are well written! I had some things to point out in both, code and manuscript. I'd consider them as minor improvements to the quality of the contribution, as the validity of your main claims are assured nevertheless. Please find my review below @rsankar9 @benoit-girard .
> ## Review Summary
>
>     * Full replication | Partial replication | Failed replication | Reproduction: I believe it's a "Partial replication", but I have two things to note here @benoit-girard:
>
>       * The authors mention in their manuscript's section "2 Comparison with Python Re-implementation" that they had access to the Matlab code of the original publication's authors. This essentially introduces a bias to the replication, which became evident by that the authors made modifications to their Python code based on how the Matlab code differs from methods described in the original publication. So I believe without the Matlab code at hand it would've been a "Failed replication" due to the seemingly missing information in the original publication. The authors consider their partial replication being a "Re-implementation", which sort of reflects this circumstance and very well point out the differences they discovered between the original paper and the Matlab code. I would consider it therefore a successful partial replication, but wanted to point out this circumstance nevertheless.
>       * Is there a way to denote the difference in replications between full and partial in the article's tag (i.e. [Re] vs. [Re\] or something)? As so far the manucsript is tagged as a "Replication" besides replicating not all experiments (i.e. "Partial replication").
>
>

I agree with this assessment. It's a partial replication due to two reasons.

1. I have replicated the three principal tasks (along with some variants to test robustness and scalability) mentioned in the paper. However, the authors proceed to test the model further with some other variants that I haven't included.
2. I agree that referring to the author's code does introduce a bias, especially since I've strived to have a close Python adaptation of the MATLAB scripts. However, I do believe it is possible to replicate the model successfully without the author scripts. The only crucial missing information in the paper, when I attempted to replicate without the MATLAB scripts, were the learning rates for the RMHL weight update.

>     * Licensing: BSD-2-clause license for both Matlab code (by Pyle and Roberts) and Python code (by the contributors of this submission)
>
>     * Reproducibility of the replication: I was able to run the Python code without any problems. The Matlab code did not work under GNU Octave 6.2.0. As Matlab is closed-source I'm not sure what the directive would be here either way @benoit-girard?
>
>     * Clarity of the code: Overall well structured and documented code.
>
>     * Clarity and completeness of the accompanying article: Concise and easy to read, captures the main claims and contributions of the original publication and points out robustness issues encountered in certain experiments as well as a possible solution. Some points could be alittle bit clearer (see my comments below).
>
>
> ## Remarks w.r.t. the manuscript
>
>     * Layout: reading through the references is a little bit complicated, since LaTeX obviously cluttered some graphics inbetween the list of references (Figs. 5, 6). The figures should be placed before the "References" section starts.

>     * References: Reference 1 is missing the journal name and is formatted differently than refs 2-6 (title in bold instead of journal, no quotation marks around title etc.). Some journal names are in title case others in sentence case.

Fixed.

>
>     * Framework:
>
>       * "This allows for partially unsupervised learning" (p.2) - what is  meant by "partially unspervised learning"?

>       * The article should make for a self-contained read without reiterating every detail from the original publication. For my taste, this would include stating the complete model equations, i.e. reservior equations, computation of `z_1`, `z_2` from `r`, and mentioning the numerical solver used to simulate the system of differential equations, so that the reader knows about the meaning of the weights updated by eq.1-3.
>       * While FORCE and RMHL acronyms are epxplained, this was omitted for SUPERTREX (p.2).

* I have removed the sentence with the incorrect phrasing of "partially unsupervised learning".
* I've included the model equations for the three algorithms. No specific numerical solver has been used to simulate the differential equations, since the reservoir dynamics and weights are merely updated iteratively, at each discrete timestep.
* I've added the SUPERTREX acronym explanation.

>     * Task:
>
>       * "each tasks uses [...] multi-segmented arm" (p.2) - some tasks are based solely on the pen's position. Admittedly, one could view that as a 0-segment arm condition, though.
>       * Concise language: "A non-linear inverse transformation would be required" (p.2) - transformation of what? having read the original work this is clear to me, but might not be clear for someone who doesn't know/forgot about the details in the original work.
>       * Reading this section as well as "Task 1" and "Task 2" it might seem to the reader as if RMHL and SUPERTREX are learned with non-scalar error values for Task 1, as for Task 1 it is solely noted that "known target" output is used for training while for Task 2 it is explicitely stated that the error is a scalar. Indeed, for Task 1 the error (to the reward-modulated learner) is as well of scalar type. Critically, though the distinction comes from how the scalar is computed: output and target domain match in Task 1 - distance of the pen's position is indeed the target value of the output - the error couples linearly to the correct solution's output values -, while for Task 2 and 3 the output domain is angular values, but the target domain is the end effector's position - the error is non-linearly coupled to the correct solution's output values.

* I agree that one task (Task 1) is based on the pen's position alone without any manipulation of the arm segments. I have replaced this misleading statement. The paragraph now reads "three motor tasks, with increasing difficulty...the target of each task is to learn to produce a given spatio‐temporal signal, under different constraints...Task 2 and Task 3 use the paradigm of exploration by a multi‐segmented arm, pivoted at a point."
* This sentence has now been replaced with "A non-linear inverse transformation of the trajectory would be required to compute the desired angles between the arm segments."
* I agree with your assessment here. To address this I have replaced the phase "scalar error signal" with "indirect error signal", as I believe the  main objective here is to emphasize on the increased complexity of the task.

>     * Task 1,2,3:
>
>       * The explanation should note, that during testing the feedback is provided via teacher forcing (cf. the original publication), since it was shown to be of quite important impact for the model performance.

* I've included this information in section 1.2. "It is also worth noting that the authors use teacher-forcing in the testing phase, which considerably improves the model performance by limiting the dependence on the stability of the learned solution (refer to Section "State information provides stability of learned output" in ~\cite{pyle2019})".

>     * Task 3:
>
>       * "We compare the performance of the MATLAB scripts and our Python adaptation for the three algorithms on Task 2, with the results presented in the article." - This sentence isn't clear to me. I believe it should say "Task 3" instead of "Task 2" here? Also describing your Python (re-)implementation as an "adaptation" might be a little bit confusing in the light that you later on indeed modify (i.e. adapt) your Python implementation to increase robustness.
>       * I don't believe the text accurately reflects the degree to which the results are in line with the original publication here. While the later sentences acknowledge a deviation from the target contour, the summarizing introduction ("The MATLAB scripts and the Python re‐implementation are able to successfully reproduce the results presented in the paper with the default seed as well as with the 10 arbitrary seeds for the RMHL algorithm.") raises the impression, that the results in the orginial publication look about the same. Indeed, they don't, which is also why you changed the seeds (which nevertheless do not make up for the complete discrepancy w.r.t. the original publication). See also my comments w.r.t. "Results" in "Remarks w.r.t. the implementation".

* You're right. I've fixed the typo. Also, to address this ambiguity about the terms adaptation and re-implementation, I've added a terminology explanation in the beginning. We use the term Python adaptation, for the scripts wherein we have presented a close adaptation of the original MATLAB scripts, in Python. We use the term re-implementation for our Python scripts, which include the modifications and some simplifications to the author scripts (and hence, is not a strict adaptation, but indeed, a re-implementation).  Hopefully, this resolves the issue. I'm open to suggestions for a better terminology.

>     * Modification:
>
>       * Figure 4 is never referenced, but I feel like it should be noted somewhere within this section?
>       * "We observe that RMHL performance is comparable to the original Task 2, [...]" - this is nowhere shown, right? So maybe add a "(not shown)".
>       * The "two minor alterations" you stated refer to Tasks 1-3 throughout the section. To me it is not entirely clear whether the results of the privous sections have been generated with or without these alterations. From the structure of the manuscript and the labelling of the figures I would expect that the alterations mentioned are only applied for experiments conducted within "3 Modification". But, on the other hand, why would one state then how these alterations relate to Tasks 1-3?
>       * It is great to have the two alterations stated as a bullet point list. What might be a little confusing is the reiteration of the two points (with an added interpretation of them) in the subsequent paragraph as by using "also" in "We also increase the error threshold [...]" it seems, on a first look, like here comes an additional alteration on top of the ones from the bullet point list, while it is indeed the second item of the list.
>       * "[...] the model is able to perform well with up to 50 arm segments (Figure 5)." - Again, I believe "well" is a little bit optimistic for some of the results. For e.g. 10 arm segments (Fig. 5) I wouldn't use the word "well", while for e.g. 6 arm segments this seems to be justified.

* I've included the reference to Fig 4.
* Added 'not shown' in the caption.
* I've altered the paragraph in the manuscript to increase clarity. To answer your question here, the first three figures have been generated _without_ these alterations. However, these alterations do, indeed, refer to all three tasks. In the interest of uniformity, the new scripts for the three algorithms stay the same for all three tasks. So, I re-simulate the three tasks with the modified Python re-implementation. (To verify the performance in the existing satisfactory cases doesn't decrease.) I haven't shown the figures for Task 1 and Task 2, as there is no significant improvement. I have now included a deviation metric in this paragraph to make this more clear. I have also included the deviation metric to show the improvement in Task 3 performance. I haven't include the figures as I believe it doesn't convey any additional information to the deviation metric and the previous figures. I have only included the figures for the Task 2 variant to showcase the improved robustness and scalability of the model.
=> Tldr: I state these alterations in relation to Task 1-3, as _with_ the alterations, the model has equal or better performance for all three tasks.
* I've rephrased the reasoning for why the alterations improve performance as follows: "Alteration \#1, by including a compensation factor for the change in number of arm segments, prevents the weights from increasing exponentially, and lets the simulation proceed in a meaningful manner. Alteration \#2, by increasing the error threshold governing the transfer of information to the mastery pathway, makes the model more tolerant of fluctuations, while continuing to explore and learn a good solution."
* This is a very valid point. I've included a deviation metric now, both as a summary statistic in a table, and indicated them on the MSE plots. So this should provide a more quantitative measure about the quality of performance. Also, I've rephrased the sentence using "satisfactory" instead of "well". This is based on the threshold I set for satisfactory vs unsatisfactory performance. Also the intention here is to show that the simulation proceeds without an exponential increase in weights. I have edited the paragraph to make this more clear.

>     * Figures:
>
>       * No indication for the temporal extent of the plots is given (x-axis) (cf. original publication).

I have added this information in the caption now.

>       * I believe the "Original" butterfly plots are taken from the original publication. If so, this needs to be stated (and potential permission issues would have to be clarified with the holder of their copyrights).

It's true. I will contact the authors and keep you posted. Meanwhile, I have added a footnote indicating these plots have been re-used from the original paper.

>       * The "Original" butterfly plot looks squeezed w.r.t. the horizontal extent. Is this just a visual feat, or as well corresponding to differencees in the underlying implementation?

I use the aspect ratio of 1:1. Also, the scripts of the authors seem to use the aspect ratio 1:1. So, the plots, perhaps, from the paper look squeezed due to the way the journal prints it.

>       * For viewing the document digitally, it might be beneficial to use vector graphic plots instead of pixelated information. Note the difference when zooming in into the figures. Just in case this is easily realizable and doesn't require too much effort.

All the figures are in "eps" format now.

>       * The different plots within a figure aren't drawn to the same scale (not the difference in "MATLAB" and "Python" red butterfly outlines, or the vertical scaling of e.g. x and y axes), which makes it harder to visually judge relative perofrmance. Same goes for the horizontal scaling of temporal axes where weight changes are depcited (i.e. the gray vertical bars don't align with them of the angular evolutions).

I have now aligned the scales for corresponding plots.

>       * No plots for the evolution of the performance metric (distance from target) and weights (norm of the weight matrix) are shown. This would aid the comparison with the original publication and, I believe, would make the case for a successful replication even stronger.

I have added the MSE subfigures to all tasks, and the W\_norm plot to all tasks except task 3, due to space/legibility constraints.

>       * Some plots (in Figs. 4,5) have scales for their vertical axis plotted, which is great. Unfortunately, the ticks of these scales are too small and squeezed to be interpretable.

Improved.

>       * For some of the temporal plots a slighlty increased linewidth would helpful.

Fixed for all plots.

>       * Up to taste: Using the same color tone (same blue) for "MATLAB" and "Python" temporal evolution plots would be pleasant.

Fixed for all plots.

>       * Figure 1: It seems like "MATLAB" and "Original" butterfly plots are consisiting of one period of drawing the shape (blue line), while "Python" seems to have multiple periods visualized. Is this the case, or are the different periods just aligning perfectly?

* Good catch. All three have multiple periods visualised (all 5 periods of the test phase). Two explanations for the Python one seeming more chaotic. One (minor), the Python one was sometimes not using the same timescale for the low pass filter used for the visualisation (Fixed that now). Two (major), the test phase uses teacher forcing. For task 1, the authors use the feedback of target output. For task 2 and 3, they use the feedback of the model's output from the training phase (since the target output is unknown). However, in my Python scripts, I had used the second type of feedback even in task 1 (in my attempt to maintain uniformity). This lead to the decreased performance in comparison. I've rectified that now. And the results should look comparable now.

>       * Figure 2:
>
>         * For the plots from the original article there seems to be some vertical gray artifact line to the left of each plot (it vanishes and appears w.r.t. my reader's zoom factor).

I was unable to reproduce this. I have re-uploaded new figures, in any case. Could you let me know if the issue persists?

>         * The caption seems to have a copy-paste-error: ""The second row shows [...] (x and y coordinates, in this case) [...]" - actually, the angles are depicted (according to the vertical axes' tag).

Fixed.

>       * Figure 3:
>
>         * For the plots from the original article there seems to be some vertical gray artifact line to the left of each plot (it vanishes and appears w.r.t. my reader's zoom factor).

I was unable to reproduce this. I have re-uploaded new figures, in any case. Could you let me know if the issue persists?

>         * Suplot "(a)"/"(c)": "The target time-series is imitated well by the model during the training phase [...]" - this information is not shown and should be denoted as such (i.e. "(not shown)").

Fixed.

>         * Subplot "(b)": The y-axes of the plots of evolutions of the angles are somehow having an additional, squashed tag for each angle on to their left. These are probably due to how the plots have been generated.

Fixed.

>         * Subplot "(c)": I'm not sure whether "[...] with slight divergences [...]" does the subplots do justice. Qualitatively, the results are still of butterfly shape, but quantitatively, I believe, the performance metric will be quite impacted during some phases of each period (not shown).

I've included a diversion metric now, both in the caption and, as an aggregate sumamry in a table. This should hopefully justify the statement.

>       * Figure 4:
>
>         * The differently colored lines of the evolution of the matrices' norm isn't described (i.e. green = exploration, purple = mastery pathways / `W_1` and `W_2`).

Fixed.

>
>     * Discussion:
>
>       * "Only two necessary details were missing, [...]" - if they have been provided by the original authors upon request, it would be worth to mention it. Or have they been found by trial-and-error?

* I mention these two details are not "described in the paper." However, they are available in the MATLAB scripts. (Before I had access to these scripts, I was using a slightly different value found by trial-and-error).

>       * "[...] the inclusion of a crucial learning rate of 0.0005 [...]" - can you explain which learning rate this refers to? If I read the original publication's method section correctly, then they state a learning rate of k=0.5 for Tasks 1 and 2 and k=0.9 for Task 3 (p. 1457).

* I use the term learning rate, as this is the variable name that the authors use in their script. This is different from the learning rate (k) that they use for SUPERTREX. This is used to control the weight update of the RMHL pathway.

[…]
