### 2021-07-25 #50 @rgutzen
The revised article looks very good to me, and I'm satisfied with how all the review comments were addressed.
Especially, the addition of the error traces is very helpful as well as the quantification of the results in the overview tables.
For me, there is no additional review round required.
I just noted down a few minor things for the revised version, whose fixing I'd consider however optional.

- Fig. 1 caption: I don't quite understand "X‐axis scale:‐ 1s : 1 period"

- In several figure captions you write "The horizontal grey line, in the test phase, indicates the deviation metric." which may be a bit ambiguous whether this metric is imposed or derived. To make this clearer you could write '... indicates the average MSE which is used as the deviation metric'.

- page 8, bullets: missing whitespace 'n=11)and'; additional whitespace 'FORCE algorithm.  The'

- Fig. 5: It took me a bit to understand that there is nothing missing on the left Matlab side. Maybe, you'd like to add a note or indication to the plot that the simulation broke after x seconds, to make this clearer to the reader.
