### 2021-08-13 #50 @rsankar9
@rgutzen Thank you for your comments. I've made some minor changes in response to them.

- In Figure 1, I've added a graphical marker to indicate the scale.
- On page 4, I have described the term deviation metric, as I'll be re-using it all across the paper. I've now emphasised it in bold, so that it is more clear to the reader.
- Fixed the typos.
- I've annotated the figure to make this clear.
