### 2021-11-01 #50 @rsankar9
Thanks to both the reviewers for your thorough and helpful comments. I've made some modifications to the manuscript. Hopefully it should address your concerns.

> ## Review Summary
>
>     * The article as well as the implementation have been greatly improved since my last review. As well, all my points have been addressed satisfactorily. Additional (sub-)figures and clear metrics for judging the quality of simulation outcomes make for a great, understandable read. Given the specificities of the work and the code base provided by the original paper's authors, I appreciate that the article's authors tried to kept as many parallels as possible in their implementation w.r.t to the Matlab implementation - this will help potential users with cross-checking between the implementations and fosters understanding. Furthermore, the code is nicely formatted and documented with insightful `README.md` files. Overall, I believe, this qualifies the article and code at their current state for publication. From my side, no additional round of reviewing would be required. @benoit-girard
>
>     * While reviewing article and code I nevertheless spotted some minor points, which I still want to point the authors to. In my view, these can be tackled without the need for another review. Please, view these changes as optional either way.

> ## Remarks w.r.t. the manuscript
>
>     * In Table 1 you mention `Note that for variant of task 2, the SUPERTREX statistics have been computed using only 2 simulations.`. I believe this is true for the `MATLAB` and `Python adaptation`, but probably you used all available 11 samples for the `Python re‐implementation` (and thus want to mention that)?

You're right. For the re-implementation, it uses all available samples. I've updated the caption to make the distinction.

>     * Some axis labels seem a little bit warped (stretched horizontally or compressed vertically)

I see what you mean. This is caused due to difference in aspect ratio between the original figure (produced by the scripts) and the figure displayed in paper (scaled to fit by latex). This can be solved by changing the aspect ratio / figure size when plotting the original figure. If it's ok, I would choose not to fix this as it would require regenerating all the figures. Hopefully, the figures provided along with the manuscript, in its original aspect ratio, should be sufficient for readers who would like more details.

>     * It is nice to have the scale of the horizontal axis included in Fig. 1, bottom right. Probably the other figures would profit from a similar indication.

I've included it in the other figures now.

>     * Fig.4c subcaption could profit from mentioning which implementation has been used to produce the result.

The implementation used is the same as mentioned on top of the columns. I've added this information to the caption of Fig 4c to be more clear. I believe your doubt arises due to the difference in seeds. The reason for this is that since the default seed did not yeild the desirable result (as shown in Fig 4b), I've shown a sample from one of the 10 extra runs. These runs use randomly chosen seeds for the generator, and hence, do not match.

>     * Fig.7 caption last sentence: "Using the MATLAB scripts, the readout weights increase uncontrollably rendering the model unable to learn." I'm not completely sure here, but probably it is a left-over from another figure? Seemingly, it doesn't fit the context of Fig.7.

Good catch. Yes, it is a carry over mistake. Thanks for pointing out.

>
>     * Minor typos etc. I spotted during reading (optional to be adapted):
>
>       * p.1: singular-plural "Most existing algorithms are built on fully supervised learning
>         rules (e.g. FORCE [2]), which limits its potential" - should probably be "their potential"
>       * p.5 "The mean deviation , for" - there's a whitespace in front of the comma
>       * p.5 "[...] for both the original scripts (0.168±0.038;
>         n=11)and" - there's a whitespace missing after the ")"
>       * p.5 "Figure 1,2;" - there's a whitespace missing after the comma
>       * p.12 "On the other hand, simulations of [...] seeds was able to" - should be plural: "[...] were able"

Fixed the minor typos.

>     * There are some overflow problems at the end of lines (optional to be adapted):
>
>       * p.10 tables overflow onto the page's right margin and also some columns overflow (e.g. "Task" column in Table 1)
>       * p.12 "11; Modified Python re‐implementation: 0.140 ± 0.071, n=11)." overflow onto the page's right margin
>       * p.16 "Python re-implementation" overflow onto the page's right margin

Fixed the overflow issues.

> ###     * Answers to previous discussion points:
>
>
> > * I believe the "Original" butterfly plots are taken from the original publication. If so, this needs to be stated (and potential permission issues would have to be clarified with the holder of their copyrights).
> >
> >   * It's true. I will contact the authors and keep you posted. Meanwhile, I have added a footnote indicating these plots have been re-used from the original paper.
>
> Thanks for adding the footnote. Just to make sure: Has there been an answer from the authors by now? (no action in the document required I guess)

We do have a response from the authors. They agree to for both their codes and their figures to be used with the appropriate license (included in the repository for MATLAB scripts) and citation (included in the I've filled a request form (as per https://direct.mit.edu/journals/pages/rights-permissions) and am awaiting their response.

> > * For the plots from the original article there seems to be some vertical gray artifact line to the left of each plot (it vanishes and appears w.r.t. my reader's zoom factor).
> >
> >   * I was unable to reproduce this. I have re-uploaded new figures, in any case. Could you let me know if the issue persists?
>
> Seemingly the issue still persist for me, but I cannot tell why. Here is a screenshot from Fig. 3, b) at zoom factor 100% using Adobe Acrobat Reader on Windows 10 (although the issue is as well there for some of the other figures): [link to the screenshot](https://imgur.com/L7HSnlI) Still, if this is just an issue on my end, then there's no action required here.

I remain unable to reproduce this issue. However, I use a different OS and pdf viewer. If it doesn't disrupt the understability of the document, I reckon we let it be. I'm open to solutions, if someone else is able to reproduce this error.

> ## Remarks w.r.t. the implementation
>
> In order to be able to run your code I needed to (would be nice to have those adapted):
>
>     * comment out in Python adaptation/ModelFORCE.py:101,116 usage of `J_mat` as the file `var.mat` doesn't exist
>
>     * the requirement for `json==2.0.9` throws an error (at least on Ubuntu https://stackoverflow.com/questions/41466431/pip-install-json-fails-on-ubuntu) and `pip install -r requirements.txt` breaks, but everything is still functional if commented out in the respective `requirements.txt` files
>
>     * the requirement `pandas` is missing in the respective `requirements.txt` files

The corresponding changes have been made.
