### 2021-11-03 #50 @rsankar9
> > > * I believe the "Original" butterfly plots are taken from the original publication. If so, this needs to be stated (and potential permission issues would have to be clarified with the holder of their copyrights).
> > >
> > >   * It's true. I will contact the authors and keep you posted. Meanwhile, I have added a footnote indicating these plots have been re-used from the original paper.
> >
> >
> > Thanks for adding the footnote. Just to make sure: Has there been an answer from the authors by now? (no action in the document required I guess)
>
> We do have a response from the authors. They agree to for both their codes and their figures to be used with the appropriate license (included in the repository for MATLAB scripts) and citation (included in the I've filled a request form (as per https://direct.mit.edu/journals/pages/rights-permissions) and am awaiting their response.

Just a quick update regarding this.
The journal has granted nonexclusive permission to reprint the figures and equations from “A Reservoir Computing Model of Reward Modulated Motor Learning and Automaticity” in this article, provided that the original Neural Computation article is cited appropriately.
