+++ Code/MATLAB implementation/Task2_ST_Seg3_Var.m
@@ -1,0 +1,354 @@
+% Task 2 - SUPERTREX  - 3 Segments - Variable lengths 
+% This script was provided by the Ryan Pyle and Robert Rosenbaum
+% Slight modifications were made by Remya Sankar for ease of use and have been marked
+% within the script. No changes were done to the simulation parameters.
+%
+% This script simulates the SUPERTREX algorithm on a variant of Task 2 as presented in
+% Pyle, R. and Rosenbaum, R., 2019.
+% A reservoir computing model of reward-modulated motor learning and automaticity.
+% Neural computation, 31(7), pp.1430-1461.
+% Arm:- [1.8, 1.2, 0.6]
+
+   
+function y = Task2_ST_Seg3_Var(triallen, res_folder, green, red, orange, purple, grey, rseed) % Alteration #1
+
+    rng(rseed);                                                                           % Alteration #2
+    
+    res_path = res_folder+num2str(rseed);                                               % Alteration #4
+    mkdir(res_path);                                                                         % Alteration #2
+
+    dt = .2;
+    tau = 10;
+    Triallen = triallen;                                                                % Alteration #3: 1e4;
+    numTrial = 15; % 10 train, 5 test
+    numTrain = 10;
+    alpha = 2.5e-2;
+    beta = 1e-2;
+    N = 1000; % reservoir size
+    p = .1; % connection probability
+    lambda = 1.5; % eigenradius, >1 means unstable
+    Jvar = 1/(p*N); % standard deviation
+    J = sprandn(N,N,p) * lambda * sqrt(Jvar); % connectivity
+    J = full(J);
+    save(res_path + '/var', 'J');
+    gamma = 10;
+    P = eye(N)/gamma; % FORCE inverse correlation estimate initialization
+
+    % Task (Target)
+    butt = @(theta) 9 - sin(theta) + 2*sin(3*theta) + 2*sin(5*theta) - sin(7*theta) + 3*cos(2*theta) - 2*cos(4*theta);
+    numT = Triallen/dt;
+    thetas = linspace(0,2*pi,numT);
+    for i = 1:length(thetas)
+        xout(i) = butt(thetas(i))*cos(thetas(i))/14.4734; % 14.4734 is max r
+        yout(i) = butt(thetas(i))*sin(thetas(i))/14.4734;
+        sout(i) = i/length(thetas);
+        sout2(i) = cos(thetas(i));
+        sout3(i) = sin(thetas(i));
+    end
+    NRO = 3; % Dimensionality of output
+    f = zeros(2,numT); % target function
+    f(1,:) = xout;
+    f(2,:) = yout;
+    f = repmat(f,1,numTrial);
+    % Double State Feedback (circle)
+    fs = zeros(NRO,numT); % target function
+    fs(1,:) = sout2;
+    fs(2,:) = sout3;
+    fs = repmat(fs,1,numTrial);
+    
+    
+    % Position of joints from output (angles) [3 segments]
+    Poseqn = @(x) [0+ sin(x(1)*pi)*1.8 + sin((x(1)+x(2))*pi)*1.2 + sin((x(1)+x(2)+x(3))*pi)*.6;
[…]

[…]
