### 2021-11-22 #50 @benoit-girard
@rsankar9 thanks for the very last updates on the paper. Based on the positive feedback provided by the reviewers, @rgutzen @schmidDan the paper is accepted for publication in ReScience C.
I will start now the tedious editing process, I will soon ask you to recompile your paper with updated data.

@rgutzen @schmidDan : thanks for the time and expertise invested in this review process!
