### 2021-11-22 65c57ad
Updated metadata.yaml with edition step data (editor, reviewers, dates and DOI)
