### 2021-11-22 #50 @benoit-girard
@rsankar9 : I updated the metadata.yaml file, I made a Pull Request on your repository for that. I need you to recompile your pdf with the following data: updated editor & reviewer information, and dates of acceptance and publication. After that I will finalize publication.
