### 2021-11-29 #50 @rsankar9
@benoit-girard Thank you.
I have merged your PR. The PDF has the updated information. Feel free to let me know if there's anything else to be done.
