### 2021-12-03 #50 @benoit-girard
Ok, I see that the process has been finished (by @rougier ?), and that the paper is officially published and referenced on the journal website.
Congratulations!
